data_rest_api/requirements.txt
mysqlclient==1.3.10
Django==1.9.10
django-extensions==1.8.0
python-dateutil>=2.1
djangorestframework==3.6.3
Pygments==2.2.0
httpie==0.9.9
